third_party/blink/web_tests/platform/generic/fast/forms/numeric-input-name-expected.txt

Check form element indexed iteration (there should be no identical lines).

field_000=Field_000
0=0
-1=-1
method=Method
reset=Reset
focus=Focus
children=Children
